敵出現時に演出（ディゾルブ、ライト）を追加。プレイヤー移動速度変更。ライトマネージャにimgui

+++ Application/Character/CharacterState.h
@@ -9,7 +9,7 @@ class Character;
 class CharacterState
 {
 protected:
-	int32_t timer_ = 0;
+	float timer_ = 0;
 	RaycastHit info_;
 
 public:
